Implement FindClientCommand and made FindCommand abstract
