+++ src/main/java/seedu/address/logic/commands/client/FindClientCommand.java
@@ -1,0 +1,57 @@
+package seedu.address.logic.commands.client;
+
+import static java.util.Objects.requireNonNull;
+
+import seedu.address.commons.util.ToStringBuilder;
+import seedu.address.logic.Messages;
+import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.FindCommand;
+import seedu.address.model.Model;
+import seedu.address.model.client.NameContainsKeywordsPredicate;
+
+/**
+ * Finds and lists all clients in address book whose name contains any of the argument keywords.
+ * Keyword matching is case insensitive.
+ */
+public class FindClientCommand extends FindCommand {
+
+    public static final String COMMAND_WORD = "find_client";
+
+    public static final String MESSAGE_USAGE = COMMAND_WORD + ": Finds all clients whose names contain any of "
+            + "the specified keywords (case-insensitive) and displays them as a list with index numbers.\n"
+            + "Parameters: KEYWORD [MORE_KEYWORDS]...\n"
+            + "Example: " + COMMAND_WORD + " alice bob charlie";
+
+    public FindClientCommand(NameContainsKeywordsPredicate predicate) {
+        super(predicate);
+    }
+
+    @Override
+    public CommandResult execute(Model model) {
+        requireNonNull(model);
+        model.updateFilteredClientList(predicate);
+        return new CommandResult(
+                String.format(Messages.MESSAGE_CLIENTS_LISTED_OVERVIEW, model.getFilteredClientList().size()));
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        if (other == this) {
+            return true;
+        }
+
+        // instanceof handles nulls
+        if (!(other instanceof FindClientCommand otherFindCommand)) {
+            return false;
+        }
+
+        return predicate.equals(otherFindCommand.predicate);
+    }
+
+    @Override
+    public String toString() {
+        return new ToStringBuilder(this)
+                .add("predicate", predicate)
+                .toString();
+    }
+}
